==> 00_testField.ipynb <==
### Experiment field for new code, you can see the result rightaway

import numpy as np
import cartopy.crs as ccrs
import pickle
from helper import *
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.colors as colors
import matplotlib.dates as mdates
import math


with open('test.pkl', 'rb') as f:
    groupedYears = pickle.load(f)


i = 1

# density map code:
def simpleDF(df):
    df_with_counts = df.copy()
    df_with_counts['year'] = df_with_counts['date'].apply(lambda d: d.year)
    df_with_counts = (
        df_with_counts.groupby(["year", "systemNum", "profileNum"])
        .size()
        .reset_index(name='count')
        )
    return df_with_counts

group_zero = simpleDF(groupedYears[i])
print(len(groupedYears[i]))
print(np.sum(group_zero["count"]))
print(len(group_zero["count"]))
# print(group_zero.head())
print(group_zero[group_zero["systemNum"] == 6].head())
print(len(group_zero[group_zero["systemNum"] == 6]))

# Group and count
def simpleDF(df):
    df_with_counts = df.copy()
    df_with_counts['year'] = df_with_counts['date'].apply(lambda d: d.year)
    df_with_counts = (
        df_with_counts.groupby(["lat", "lon", "year", "systemNum", "profileNum"])
        .size()
        .reset_index(name='count')
        )
    return df_with_counts

group_zero = simpleDF(groupedYears[i])
# print(group_zero)
# print(group_zero[group_zero["year"]==2004])
print(len(group_zero[group_zero["year"]==2005]))
print(len(group_zero[group_zero["year"]==2006]))
print(len(group_zero[group_zero["year"]==2007]))


check_unique = (
    groupedYears[i].groupby(['date', 'profileNum',"systemNum"])[['lat', 'lon']]
    .nunique()
    .reset_index()
)
print(check_unique[check_unique["lat"] > 1])


df = groupedYears[i]
check_unique = (
    df.groupby(['date', 'profileNum', "systemNum"])[['lat', 'lon']]
    .nunique()
    .reset_index()
)
print(check_unique)

# Only those with more than 1 unique lat or lon
not_unique = check_unique[(check_unique['lat'] > 1) | (check_unique['lon'] > 1)]

for _, row in not_unique.iterrows():
    date = row['date']
    profile = row['profileNum']

    # Filter original df for this pair
    subset = df[(df['date'] == date) & (df['profileNum'] == profile)]

    # Get distinct lat/lon values
    unique_locs = subset[['lat', 'lon',"systemNum"]].drop_duplicates()

    print(f"\n(date: {date}, profileNum: {profile}) has multiple locations:")
    print(unique_locs)
# add ITP system number to the inital dataset
